% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{spark.sas7bdat-package}
\alias{spark.sas7bdat-package}
\title{Read in SAS datasets (.sas7bdat files) into Spark}
\description{
'spark.sas7bdat' uses the spark-sas7bdat Spark package to load SAS dataset in parallel in Spark.
}

